server: Use a boost bimap to store subscriptions
Subscriptions are stored in a complex container representing a many
to many relationship between connection handles and subscription topics.
The container relies on the boost::bimap to create a bidirectional link
between two "left" and "right" std::multimap representing the one-to-many
associations between handles and strings repsectively.
A specific ordering function is provided that allows comparison between
connection/topic pairs.

Thanks to this new container, removing pending subscriptions when
closing a connection was a piece of cake.